not bad
2.We all know that the additive primaries are red, green, and blue. Now, write a Python program that will take a color sequence as
a string from the user where R represents Red, G represents Green and B represents Blue. The program should print the choice of
colors that is actually a tuple containing the sub-tuples as (color_name, color_frequency) only if the color_frequency for that color
is at least one in the given color sequence. [You are not allowed to use count() function here.]