okay
 Assume, you have been given a dictionary named dict_1. The values of the dictionary will be a list or a tuple depending on the  case of the dictionary key. If the key is in lower case, then the value will be in list format. If the key is in upper case, then the value  will be in tuple format.
Now, write a Python program that creates a new dictionary named “dict_primes” that contains only the prime numbers in the  value. Then, finally print the dictionary dict_primes.
=============================================================================
Given dictionary1:
{"a":[5,2,55,17],"P":(11,121,222),"B":(37,53,71),"c":[45,92,50]}
Sample Output 1:
dict_primes = {'a': [5, 2, 17], 'P': (11,), 'B': (37, 53, 71), 'c': []}
Explanation1:
Among the elements of [5, 2, 55, 17], the prime numbers are 5, 2, 17. So, they are added to the value. The key a is in lowercase, so the value will be a list.
==========================================
Among the elements of (11, 121, 222), the prime number is 11. So, it is added to the value.
The key P is in uppercase, so the value will be a tuple.
==========================================
Among the elements of (37, 53, 71), the prime numbers are 37, 53, 71. So, they are added to the value. The key B is in uppercase, so the value will be a tuple.
==========================================
Among the elements of [45, 92, 50], none are prime numbers. So, none of the numbers are added to the value. The key c is in lowercase, so the value will be a list.
=============================================================================
Given dictionary2:
dict_1 = {"N":(4,9,3),"k":[95,37,197],"F":(32,5,97),"s":[31,94,55]}
Sample Output 2:
dict_primes = {'N': (3,), 'k': [37, 197], 'F': (5, 97), 's': [31]}
Explanation2:
Among the elements of (4, 9, 3), the prime number is 3. So, it is added to the value.
The key N is in uppercase, so the value will be a tuple.
==========================================
Among the elements of [95, 37, 197], the prime numbers are 37, 197. So, they are added to the value. The key k is in lowercase, so the value will be a list.
==========================================
Among the elements of (32, 5, 97), the prime numbers are 5, 97. So, they are added to the value.
The key F is in uppercase, so the value will be a tuple.
==========================================
Among the elements of [31, 94, 55], the prime number is 31. So, it is added to the value.
The key s is in lowercase, so the value will be a list.