pretty nice
Two words are anagrams if they contain all of the same letters, but in a different order. For example, “evil” and “live” are anagrams because each contains one “e”, one “i”, one “l”, and one “v”.
Write a program that reads two strings from the user and determines whether or not they are anagrams. Use a dictionary to solve the problem