good one
Suppose you have a sitting arrangement of the students who will give the final exam in a list. The list contains the last two digits of their student ID as given below.

sitting_list = [10,30,20,70,11,15,22,16,58,100,12,56,70,80]

Now you want to organize the sitting arrangement of the students in your own way. You decide to sort all the students in the even indices of the list in ascending order and all the students in the odd indices of the list in descending order.

So, write a python program that organizes the list for you in this way.

Sample Output
[10, 100, 11, 80, 12, 70, 20, 56, 22, 30, 58, 16, 70, 15]