jjj
3. Write a function in Python that will take a space separated string text as input from the user and
returns a dictionary having the unique words as the keys and their frequency in the given text as the
values in a sorted order(ascending) according to the frequencies.
==================================================
Hints (1):
Frequencies of the words can be easily counted with built-in function count().
Hints (2):
Sorting according to frequencies can be done in two approaches.
Approach (1): Use the built-in sorted() function. Here, for the key, you need to call a custom-made
function that will return frequencies for each word. So, the sorted() function will return values sorted
according to the frequencies.
Approach (2): Make two lists. One with the keys and other one with the frequencies. For a particular
keys, values should be found in the same index of the frequencies list. Then similar to the
Assignment07 Task-5. Sort the values of frequencies and in time of swapping, swap BOTH keys &
frequencies. You can use any one of the sorting algorithms taught (Bubble sort or Selection sort)
==================================================
Sample Input1:
"go there come and go here and there go care"

Function Call1:
function_name("go there come and go here and there go care")

Sample Output1:
{'come': 1, 'here': 1, 'care': 1, 'there': 2, 'and': 2, 'go': 3}