well good practice of error handling
Read a file from the user which contains the name and salary of employees. Convert that to a dictionary where the key is the name of each employee and value is the salary of that employee. Then find the total cost for the company that it needs to pay through employees’ salary. Print the dictionary and sum if no exception occurs.

Exception to handle:

check whether file is actually there
check whether the salary in each line is a number or not.
raise error if there are extra numbers or names
=============================

Example 1:
Input:
Saadat 100
Farzad 200
Atef 300
Output:
dict={Saadat : 100, Farzad : 200, Atef : 300}
Sum = 600

=================

Example 2:
Input:
Saadat 100
Farzad 200
Atef 300
John
Output:
Extra Statement

Example 3:
Input:
Saadat 100
Farzad 200
Atef john
Output:
Salary cannot be string