okay
Write a python program that takes a string as an input from the user. If the sum of the first half of the input is equal to the sum of the last half of the input, then it prints “You won the lottery!”, otherwise it prints “You lost.”

The input can be digits only (0-9). If the length of the string is even, then you need to consider all numbers. If the length is odd, you can ignore the middle number.

[You are not allowed to use the built-in sum() and slice() for this task]

===================================================================

Sample Input 1: 37938725

Sample Output 1: You won the lottery!

Explanation 01: The first half of the input = 3+7+9+3 = 22. The last half of the input = 8+7+2+5 = 22. The summation is equal. Since the length of the string is even, we

considered all numbers.

===================================================================

Sample Input 2: 7357314

Sample Output 2: You lost.

Explanation 02: The first half of the input = 7+3+5 = 15 The last half of the input = 3+1+4 = 8 The summation is not equal. Since the length of the string is odd, we ignored the middle number, 5.