like to practice in the hard way
Write a python program that will take a string as input from the user. The input string should have a combination of BOTH the alphabets and the digits. Then, your task is to identify the digits from that input string and store those digits in a list. Finally, sort the list and print the sorted list and the sum of digits as output to the user.

Note: You are allowed to use the built-in sort() or sorted() function here.

================================

Sample Input 01: m4gt567q09y2

Sample Output 01: ['0', '2', '4', '5', '6', '7', '9']

33

Explanation 01: Here from the given input string, the digits are stored in a list. After sorting the digits, the sorted list is printed to the user. Then, the sum of the digit which is 33 is printed.

================================

Sample Input 02: 954217

Sample Output 02: There is no alphabet in the string

Explanation 02: Here from the given input string, there is no alphabet in the string. That is why it is printing “There is no alphabet in the string”

================================

Sample Input 03: abcdefgh

Sample Output 03: There is no digit in the string

Explanation 03: Here from the given input string, there is no digit in the string. That is why it is printing “There is no digit in the string”