jj
Write the python code that takes a String from the user. Then counts the number of vowels and the consonants present in that string and calculates the ratio of vowels & consonants. Finally, prints the calculated values.

[You are not allowed to use count() function]

================================

Sample Input 1: This is CSE110 2nd quiz

Sample output 1:

No of the vowels: 5

List of vowels: ['i', 'e', 'u']

No of the consonants: 10

List of consonants: ['t', 'h', 's', 'c', 'n', 'd', 'q', 'z']

The ratio of vowels & consonants is 5: 10

Explanation 01: In the input string, vowels have appeared 5 times and consonants have appeared 10 times. So, the ratio= no of vowels: no of consonants = 5:10.

================================

Sample Input 2: This is List-String quiz

Sample Output 2:

No of the vowels: 6

List of vowels: ['i', 'u']

No of the consonants: 14

List of consonants: ['t', 'h', 's', 'l', 'r', 'n', 'g', 'q', 'z']

The ratio of vowels & consonants is 6: 14

Explanation 02: In the input string, vowels have appeared 6 times and consonants have appeared 14 times. So, the ratio= no of vowels: no of consonants = 6:14.