a
Write a python program that will take a string as an input from the user. Your task is to check the ASCII value of every character. If the ASCII value of the character is even, replace that character with the previous index character of the original string; or else keep the same. For the 0th index, since there is no previous character replace it with ‘1’

[You cannot use the built-in replace() function.]

================================

Sample Input 1: Python

Sample Output 1: 1yytoo

Explanation 01: The ASCII value of 'P'=80, 'y'=121, 't'=116, 'h'=104, 'o'=111, 'n'=110. Since the ASCII of ‘P’ is even, it is replaced by ‘1’. In the same way, the ASCII of ‘t’ is even, so replaced by the previous character ‘y’. Next, the ASCII of ‘h’ is even, so replaced by the previous character ‘t’. Finally, the ASCII of ‘n’ is even, so replaced by the previous character ‘o’.

================================

Sample Input 2: grow

Sample Output 2: ggow

Explanation 02: The ASCII value of 'g'=103, 'r'=114, 'o'=111, 'w'=119. Since the ASCII of ‘r’ is even, it was replaced by the previous character ‘g’. The other character has an odd ASCII value, so they were the same