a
Write a python program that will take a string, a starting index, and an ending index as inputs from the user. The program will print the second next character from the starting index to the ending index of the string. If the ending index is not valid, then print “Invalid Index”.

================================

Sample Input 1:

bangladesh

2

7

Sample Output 1:

pincfg

Explanation1: Here the given string length is 10 and from index 2 to index 7 the characters will be “nglade”. Then it will print the second next character “pincfg”.

================================

Sample Input 2:

science

1

7

Sample Output 2:

Invalid Index

Explanation2: Here the end index is not valid. That is why it is printing Invalid Index”