aa
Write a python program that will take a sentence as an input from the user where each word is separated by a space. Your task is to print the words present in the sentence in a list format and finally print the sentence by reversing each word.

[You are NOT allowed to use built-in reverse() or reversed()]

================================

Sample Input 1:

This is the list string quiz

Sample Output 1:

['This', 'is', 'the', 'list', 'string', 'quiz']

sihT si eht tsil gnirts ziuq

Explanation 01: Here, we have 5 words in the given sentence separated by space. Then, those words have been printed in list format. Then each word had been reversed and printed in the string format. “sihT’ is the reverse of 'This' and so on.

================================

Sample Input2:

CSE110 is the first course of CSE

Sample Output 2:

['CSE110', 'is', 'the', 'first', 'course', 'of', 'CSE']

011ESC si eht tsrif esruoc fo ESC

Explanation 02: Here, we have 7 words in the given sentence separated by space. Then, those words have been printed in list format. Then each word had been reversed and printed in string format. “011ESC’ is the reverse of 'CSE110' and so on.