aa
Write a program that takes an integer as input and then constructs a
dictionary. The dictionary should have nine keys at the end, representing
numbers from 2 to 10 (inclusive). For each key, the value should be True,
if the user input is divisible by the key. Otherwise, False.

================================================

Sample Input 1:
40

Sample Output 1:
{2: True, 3: False, 4: True, 5: True, 6: False, 7: False, 8: True, 9:
False, 10: True}

Explanation1:
40 is divisible by 2, 4, 5, 8, and 10, so the values for those keys are
True. 40 is not divisible by 3, 6, 7, or 9, so the values for those keys
are False.

================================================

Sample Input 2:
11

Sample Output 2:
{2: False, 3: False, 4: False, 5: False, 6: False, 7: False, 8: False, 9:
False, 10: False}

Explanation2:
11 is not divisible by any number from 2 to 10, so the values for all keys
are False.

================================================

Sample Input 3:
15

Sample Output 3:
{2: False, 3: True, 4: False, 5: True, 6: False, 7: False, 8: False, 9:
False, 10: False}

Explanation3:
15 is divisible by 3 and 5, so the values for those keys are True. 15 is
not divisible by 2, 4, 6, 7, 8 or 9, so the values for those keys are False.