100th practice code yeee
Take a single string as an input from the user where few numbers are
separated by commas. Now, you have to check the divisibility of each of the
numbers by 2, 3, or 5.
Write a Python program that constructs a dictionary with keys “Divisible by
2”, “Divisible by 3”, “Divisible by 5”, and “None”, with values as lists of
elements satisfying the respective condition.

================================================

Sample Input1:
15, 5, 3, 12, 14, 24, 49, 42

Sample Output1:
{'Divisible by 2': [12, 14, 24, 42], 'Divisible by 3': [15, 3, 12, 24,
42], 'Divisible by 5': [15, 5], 'None': [49]}

Explanation1:
12, 14, 24, and 42 are divisible by 2, so they belong to the key named
“Divisible by 2”. 15, 3, 12, 24, and 42 are divisible by 3, so they belong
to the key named “Divisible by 3”. 15 and 5 are divisible by 5, so they
belong to the key named “Divisible by 5”. 49 is not divisible by 2,3 or 5,
so it belongs to the key named “None”