Looks tough but easy
Write a python program that asks the user for a range (a starting
number(inclusive) and an ending number(inclusive)). Make a dictionary with
all the numbers that DO NOT end with 1, 2, and 7 within that range.
Here in the dictionary, all the digits except 1, 2& 7 are the keys, and
numbers that do not end with 1, 2 & 7 are the values. Finally, print the
dictionary.

================================================

Sample Input 1:
30
50

Sample Output 1:
{0: [30, 40], 3: [33, 43], 4: [34, 44], 5: [35, 45], 6: [36, 46], 8: [38,
48], 9: [39, 49]}

================================================

Sample Input 2:
370
385

Sample Output 2:
{0: [370, 380], 3: [373, 383], 4: [374, 384], 5: [375], 6: [376], 8: [378],
9: [379]}