well
Write a python function that takes a list of numbers(integer) and a
number(integer) as arguments. Your task is to identify pairs of numbers
from the list whose ABSOLUTE difference is equal to the 2nd
argument(number). If such a pair of numbers don’t exist in the list, then
print ‘Not Possible’. Otherwise, print those pairs in a list format INSIDE
the function.

Note: You need to use the built-in function abs(number1, number2) to get
the difference between two numbers.

================================================

Function Call1:
function_name([1, 17, 22, 38, 5, 11], 5)

Sample Output1:
[[17, 22]]

Explanation1:
The difference between 17 and 22 is 5. 17- 22 = 5(absolute value). So, the
output pair is [17, 22]. Since there are no other pairs, only [17, 22] is
printed in the answer list.

================================================

Function Call2:
function_name([1, 5, 9, 15, 74], 4)

Sample Output2:
[[1, 5], [5, 9]]

Explanation2:
The difference between 1 and 5 is 4.
1- 5 = 4(absolute value). So, the output pair is [1, 5].

Again, The difference between 5 and 9 is 4.
5- 9 = 4(absolute value). So, the output pair is [5, 9].

Since, there are no other pairs, except [1, 5] and [5, 9]. So, these are
added and printed in the answer list.

================================================ Function Call3:
function_name([1, 5, 9, 13], 5)

Sample Output3:
Not Possible

Explanation3:
In this case, there is no pair that gives a difference of 5. So, “NotPossible” is printed.