aaa
Write a python function that takes a string as an argument where multiple
numbers are separated by commas. Your task is to create a dictionary that
will have the index of the list as key and summation from start to that
index as value. Finally, Return the dictionary.

================================================

Function Call1:
function_name("3,1,5,4,6,7,8")

Sample Output 1:
{0: 3, 1: 4, 2: 9, 3: 13, 4: 19, 5: 26, 6: 34}

Explanation 1:
key 0 → value 3
key 1 → value 3+1
key 2 → value 3+1+5
key 3 → value 3+1+5+4
and so on...

================================================

Function Call2:
function_name("5,1,3,7,4,8")

Sample Output 2:
{0:5, 1:6, 2:9, 3:16, 4:20, 5:28}

Explanation 2:
key 0 → value 5
key 1 → value 5+1
key 2 → value 5+1+3
and so on...