aa
Write a python function that takes a string as an argument and searches all
the characters having multiple occurrences (except the spaces) in that
string. Then, prepare and RETURN a dictionary where the keys will be the
characters and values will be the positions of those characters in the string.

================================================

Function Call1:
function_name("this is my first semester.")

Sample Output 1:
{‘t’: [0, 15, 22],‘h’:[1], ‘i’: [2, 5, 12], ‘s’:[3, 6, 14, 17, 21],
‘m’: [8,19], ‘y’: [9], ‘f’:[11], ‘r’: [13, 24], ‘e’: [18, 20, 23]}

Explanation1:
Here, the character ‘t’ has appeared in indexes 0, 15, and 22. The
character ‘h’ has appeared in index 1. The character ‘i’ has appeared in
indexes 2, 5, and 12. And so on.

================================================

Function Call2:
function_name("my student id is 010101.")

Sample Output 2:
{‘m’:[0],’y’:[1],‘s’: [3, 15], ‘t’: [4, 9],’u’:[5],’e’:[7],’n’:[8],
‘d’: [6, 12], ‘i’: [11, 14], ‘0’: [17,19,21],’1’: [18,20,22]}

Explanation2:
The character ‘m’ has appeared in index 0. The character ‘s’ has appeared
in indexes 3 and 15. The character ‘1’ has appeared in indexes 18, 20, and\

And so on.